Remade expression-hook.lisp completely. Can't say it is really based on macroexpand-dammit now.. Macroexpand-dammit had use of eval that was very annoying, and all sorts of functions i wtf-ed at. I may be overlooking something in my own code. One thing is that the expand function needs to have access to the packages in question, and the macros need to be evaluated before it's use. Expression hook still needs to track variables and their origins properly.
Expression-scan still needs to make use of that info. And expression-scan probably needs a widening of what is *in-fun*. I believe it uses
track-fun, things that are 'function-like' should derive from it, and be able to 'submit' to *in-fun* aswel. Maybe *in-fun* needs to be a stack,
too?(dunno)

autodoc2.lisp contains the new autodoc, only for functions/macros atm, and no page creation yet. The previous version seems to be too ad-hoc to, me,
especially now all expressions are scanned (somewhat)equally. The code is split up, so hopefully trackers of other expressions will be able to use them. It
 is still completely untested.