Used cl-dot to make a package graph-scanned that can make graphs of connections between scanned source code(I made a big picture with all the source code of the project.. it is useful, i did find an improvement necessary in autodoc with it.) it is already fairly customizable, but i am not sure how to incorporate it into autodocumentation  yet.
Added file tracking to scanner, specifically for packages now, perhaps it should go onto base-track. Also made
autodocumenter able to pick it up. Will likely also make an auto-asd-file maker based on it.

In order to do that, and also to improve in the future when dealing with files, added utility to handle files
better. Also moved package-stuff to libs/ because it belongs there.