Whole bunch of things. (Not sure if i noted them all here)
Most importantly, now putting output in different directories works with html, and has a good basis in general.
This is applied to the website, separating autodocumentation and website.

Added 'mention+' to autodoc instead of just using it in doc/main.lisp

Changed a bunch of macros, like if-let, when-let, curry to come from the alexandria package.

Started a little path-stuff package to help in path stuff, it is unit-tested fairly well.

Log works, but started store-log anyway, to base a log on cl-store.. Not sure if that is wise, as log was
working.. But it is _much_ neater to work with classes.

About classes, perhaps i should change def-glist to make them instead of functions, that would make stuff a lot
more 'inspectable'. I feel stupid for not realizing that before. I am sure i still don't understand 'classes are
the poor mans functions, functions are the poor mans classes' fully.

Although i think avoiding sweeping changes endangers one to get stuck in a 'local maximum', perhaps i should try
make my commits smaller otherwise :)